Testing: Refactor newly added test to work with Jest'
